Finish applying Table component to classifications
