+++ site/gatsby-site/src/pages/apps/reports.js
@@ -1,140 +1,22 @@
 import React, { useEffect, useState } from 'react';
 import AiidHelmet from '../../components/AiidHelmet';
-import { format } from 'date-fns';
 import Link from '../../components/ui/Link';
 import { faLink } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import DateRangePicker from 'react-bootstrap-daterangepicker';
 import { useTable, useFilters, usePagination, useSortBy } from 'react-table';
 import { gql, useQuery } from '@apollo/client';
 import Layout from 'components/Layout';
 import { useMenuContext } from 'contexts/MenuContext';
 import ListSkeleton from 'elements/Skeletons/List';
 import { useTranslation } from 'react-i18next';
-import { Button, Select } from 'flowbite-react';
-import Table, { DefaultColumnFilter, DefaultColumnHeader } from 'components/ui/Table';
-
-const SelectColumnFilter = ({ column: { filterValue, setFilter, preFilteredRows, id } }) => {
-  let options;
-
-  const ARRAY_COLUMNS = ['submitters'];
-
-  const filterOptionsFromArray = () => {
-    return (options = React.useMemo(() => {
-      const options = new Set();
-
-      preFilteredRows.forEach((row) => {
-        if (Array.isArray(row.values[id])) {
-          row.values[id].forEach((w) => {
-            if (w !== '') {
-              options.add(w);
-            }
-          });
-        } else {
-          if (row.values[id] !== '') {
-            options.add(row.values[id]);
-          }
-        }
-      });
-      return [...options.values()];
-    }, [id, preFilteredRows]));
-  };
-
-  if (ARRAY_COLUMNS.includes(id)) {
-    options = filterOptionsFromArray();
-  } else {
-    options = React.useMemo(() => {
-      const options = new Set();
-
-      preFilteredRows.forEach((row) => {
-        if (row.values[id] !== '') {
-          options.add(row.values[id]);
-        }
-      });
-      return [...options.values()];
-    }, [id, preFilteredRows]);
-  }
-
-  const filteredOptions = options.filter((o) => o !== '-');
-
-  return (
-    <Select
-      style={{ width: '100%' }}
-      className="mt-2"
-      value={filterValue || 'All'}
-      onChange={(e) => {
-        setFilter(e.target.value || undefined);
-      }}
-    >
-      <option value="">All</option>
-      {filteredOptions.map((option, i) => (
-        <option key={i} value={option}>
-          {option}
-        </option>
-      ))}
-    </Select>
-  );
-};
-
-const SelectDatePickerFilter = ({
-  column: { filterValue = [], preFilteredRows, setFilter, id },
-}) => {
-  const [min, max] = React.useMemo(() => {
-    let min = new Date(preFilteredRows[0]?.values[id] ?? '1970-01-01').getTime();
-
-    let max = new Date(preFilteredRows[0]?.values[id] ?? '1970-01-01').getTime();
-
-    preFilteredRows.forEach((row) => {
-      const currentDatetime = new Date(row.values[id]).getTime();
-
-      min = currentDatetime <= min ? currentDatetime : min;
-      max = currentDatetime >= max ? currentDatetime : max;
-    });
-    return [min, max];
-  }, [id, preFilteredRows]);
-
-  if (filterValue.length === 0) {
-    setFilter;
-  }
-
-  const handleApply = (event, picker) => {
-    picker.element.val(
-      format(picker.startDate.toDate(), 'yyyy-MM-dd') +
-        ' - ' +
-        format(picker.endDate.toDate(), 'yyyy-MM-dd')
-    );
-    setFilter([picker.startDate.valueOf() / 1000, picker.endDate.valueOf() / 1000]);
-  };
-
-  const handleCancel = (event, picker) => {
-    picker.element.val('');
-    setFilter([min, max]);
-  };
-
-  return (
-    <div className="flex font-normal mt-2">
-      <DateRangePicker
-        className="custom-picker"
-        onApply={handleApply}
-        onCancel={handleCancel}
-        initialSettings={{
-          showDropdowns: true,
-          autoUpdateInput: false,
-          locale: {
-            cancelLabel: 'Clear',
-          },
-        }}
-      >
-        <input
-          style={{ width: 190 }}
-          type="text"
-          className="form-control col-4 p-2"
-          defaultValue=""
-        />
-      </DateRangePicker>
-    </div>
-  );
-};
+import { Button } from 'flowbite-react';
+import Table, {
+  DefaultColumnFilter,
+  DefaultColumnHeader,
+  formatDateField,
+  SelectColumnFilter,
+  SelectDatePickerFilter,
+} from 'components/ui/Table';
 
 const query = gql`
   query ReportsQuery {
@@ -185,10 +67,6 @@ export default function Incidents(props) {
 
     setTableData(incidentDataToCellMap(data));
   }, [data]);
-
-  const formatDateField = (date) => {
-    return <>{date && <>{format(new Date(date * 1000), 'yyyy-MM-dd')}</>}</>;
-  };
 
   const cellData = React.useMemo(() => tableData, [tableData]);
